Parse csv file containing historical OHLCV data

Helper function to parse a csv file obtained from the PredictIt website, containing historical OHLCV data, into an object of class xts.

parse_historical_csv(csv_path, filename = NA)

Arguments

csv_path
Path to a csv file containing historical OHLCV data for a specific contract (Open, High, Low, Close, Volume)

filename
Optional name to give the csv file when the filepath is derived from a temporary directory

Examples

filename <- "What_will_be_the_balance_of_power_in_Congress_after_the_2020_election.csv"
csv_path <- system.file("extdata", filename, package = "predictit")
parse_historical_csv(csv_path)

#> $data
#>                     D House, D Senate D House, R Senate R House, R Senate
#> 2019-09-25 12:00:00 "0.33"            "0.52"            "0.01"           
#> 2019-09-26 12:00:00 "0.32"            "0.50"            "0.02"           
#> 2019-09-27 12:00:00 "0.31"            "0.52"            "0.01"           
#> 2019-09-28 12:00:00 "0.33"            "0.52"            "0.02"           
#> 2019-09-29 12:00:00 "0.33"            "0.51"            "0.02"           
#> 2019-09-30 12:00:00 "0.33"            "0.52"            "0.02"           
#> 2019-10-01 12:00:00 "0.33"            "0.52"            "0.02"           
#> 2019-10-02 12:00:00 "0.30"            "0.52"            "0.01"           
#> 2019-10-03 12:00:00 "0.32"            "0.52"            "0.01"           
#> 2019-10-04 12:00:00 "0.30"            "0.51"            "0.01"           
#> 2019-10-05 12:00:00 "0.31"            "0.50"            "0.01"           
#> 2019-10-06 12:00:00 "0.30"            "0.51"            "0.01"           
#> 2019-10-07 12:00:00 "0.30"            "0.50"            "0.01"           
#> 2019-10-08 12:00:00 "0.32"            "0.51"            "0.02"           
#> 2019-10-09 12:00:00 "0.30"            "0.51"            "0.01"           
#> 2019-10-10 12:00:00 "0.32"            "0.50"            "0.01"           
#> 2019-10-11 12:00:00 "0.32"            "0.50"            "0.02"           
#> 2019-10-12 12:00:00 "0.30"            "0.50"            "0.01"           
#> 2019-10-13 12:00:00 "0.30"            "0.50"            "0.01"           
#> 2019-10-14 12:00:00 "0.32"            "0.50"            "0.01"           
#> 2019-10-15 12:00:00 "0.30"            "0.50"            "0.01"           
#> 2019-10-16 12:00:00 "0.30"            "0.49"            "0.01"           
#> 2019-10-17 12:00:00 "0.32"            "0.49"            "0.01"           
#> 2019-10-18 12:00:00 "0.31"            "0.49"            "0.02"           
#> 2019-10-19 12:00:00 "0.31"            "0.48"            "0.02"           
#> 2019-10-20 12:00:00 "0.31"            "0.48"            "0.01"           
#> 2019-10-21 12:00:00 "0.32"            "0.48"            "0.01"           
#> 2019-10-22 12:00:00 "0.31"            "0.47"            "0.01"           
#> 2019-10-23 12:00:00 "0.32"            "0.49"            "0.01"           
#> 2019-10-24 12:00:00 "0.32"            "0.49"            "0.01"           
#> 2019-10-25 12:00:00 "0.32"            "0.48"            "0.02"           
#> 2019-10-26 12:00:00 "0.32"            "0.49"            "0.02"           
#> 2019-10-27 12:00:00 "0.32"            "0.48"            "0.01"           
#> 2019-10-28 12:00:00 "0.32"            "0.49"            "0.02"           
#> 2019-10-29 12:00:00 "0.33"            "0.48"            "0.01"           
#> 2019-10-30 12:00:00 "0.33"            "0.48"            "0.01"           
#> 2019-10-31 12:00:00 "0.32"            "0.48"            "0.02"           
#> 2019-11-01 12:00:00 "0.31"            "0.49"            "0.02"           
#> 2019-11-02 12:00:00 "0.32"            "0.49"            "0.01"           
#> 2019-11-03 12:00:00 "0.31"            "0.49"            "0.02"           
#> 2019-11-04 12:00:00 "0.32"            "0.48"            "0.02"           
#> 2019-11-05 12:00:00 "0.31"            "0.48"            "0.01"           
#> 2019-11-06 12:00:00 "0.31"            "0.49"            "0.01"           
#> 2019-11-07 12:00:00 "0.32"            "0.49"            "0.01"           
#> 2019-11-08 12:00:00 "0.34"            "0.48"            "0.01"           
#> 2019-11-09 12:00:00 "0.32"            "0.48"            "0.01"           
#> 2019-11-10 12:00:00 "0.31"            "0.49"            "0.01"           
#> 2019-11-11 12:00:00 "0.33"            "0.49"            "0.01"           
#> 2019-11-12 12:00:00 "0.32"            "0.48"            "0.01"           
#> 2019-11-13 12:00:00 "0.32"            "0.49"            "0.02"           
#> 2019-11-14 12:00:00 "0.32"            "0.48"            "0.02"           
#> 2019-11-15 12:00:00 "0.33"            "0.46"            "0.02"           
#> 2019-11-16 12:00:00 "0.32"            "0.46"            "0.02"           
#> 2019-11-17 12:00:00 "0.33"            "0.47"            "0.01"           
#> 2019-11-18 12:00:00 "0.33"            "0.47"            "0.01"           
#> 2019-11-19 12:00:00 "0.32"            "0.48"            "0.01"           
#> 2019-11-20 12:00:00 "0.31"            "0.47"            "0.02"           
#> 2019-11-21 12:00:00 "0.31"            "0.47"            "0.01"           
#> 2019-11-22 12:00:00 "0.32"            "0.48"            "0.01"           
#> 2019-11-23 12:00:00 "0.32"            "0.47"            "0.02"           
#> 2019-11-24 12:00:00 "0.31"            "0.48"            "0.01"           
#> 2019-11-25 12:00:00 "0.32"            "0.48"            "0.01"           
#> 2019-11-26 12:00:00 "0.31"            "0.49"            "0.01"           
#> 2019-11-27 12:00:00 "0.31"            "0.49"            "0.01"           
#> 2019-11-28 12:00:00 "0.30"            "0.49"            "0.01"           
#> 2019-11-29 12:00:00 "0.30"            "0.49"            "0.01"           
#> 2019-11-30 12:00:00 "0.31"            "0.49"            "0.01"           
#> 2019-12-01 12:00:00 "0.30"            "0.48"            "0.01"           
#> 2019-12-02 12:00:00 "0.30"            "0.49"            "0.01"           
#> 2019-12-03 12:00:00 "0.30"            "0.48"            "0.01"           
#> 2019-12-04 12:00:00 "0.30"            "0.49"            "0.01"           
#> 2019-12-05 12:00:00 "0.31"            "0.49"            "0.01"           
#> 2019-12-06 12:00:00 "0.30"            "0.49"            "0.01"           
#> 2019-12-07 12:00:00 "0.31"            "0.49"            "0.01"           
#> 2019-12-08 12:00:00 "0.31"            "0.49"            "0.02"           
#> 2019-12-09 12:00:00 "0.31"            "0.49"            "0.02"           
#> 2019-12-10 12:00:00 "0.30"            "0.49"            "0.02"           
#> 2019-12-11 12:00:00 "0.28"            "0.48"            "0.01"           
#> 2019-12-12 12:00:00 "0.27"            "0.50"            "0.02"           
#> 2019-12-13 12:00:00 "0.27"            "0.50"            "0.01"           
#> 2019-12-14 12:00:00 "0.27"            "0.49"            "0.02"           
#> 2019-12-15 12:00:00 "0.25"            "0.51"            "0.02"           
#> 2019-12-16 12:00:00 "0.24"            "0.50"            "0.02"           
#> 2019-12-17 12:00:00 "0.25"            "0.50"            "0.01"           
#> 2019-12-18 12:00:00 "0.26"            "0.51"            "0.01"           
#> 2019-12-19 12:00:00 "0.25"            "0.52"            "0.02"           
#> 2019-12-20 12:00:00 "0.25"            "0.55"            "0.01"           
#> 2019-12-21 12:00:00 "0.24"            "0.55"            "0.01"           
#> 2019-12-22 12:00:00 "0.25"            "0.57"            "0.01"           
#> 2019-12-23 12:00:00 "0.25"            "0.55"            "0.01"           
#>                     R House, D Senate
#> 2019-09-25 12:00:00 "0.23"           
#> 2019-09-26 12:00:00 "0.23"           
#> 2019-09-27 12:00:00 "0.23"           
#> 2019-09-28 12:00:00 "0.24"           
#> 2019-09-29 12:00:00 "0.24"           
#> 2019-09-30 12:00:00 "0.24"           
#> 2019-10-01 12:00:00 "0.23"           
#> 2019-10-02 12:00:00 "0.23"           
#> 2019-10-03 12:00:00 "0.23"           
#> 2019-10-04 12:00:00 "0.26"           
#> 2019-10-05 12:00:00 "0.26"           
#> 2019-10-06 12:00:00 "0.25"           
#> 2019-10-07 12:00:00 "0.25"           
#> 2019-10-08 12:00:00 "0.26"           
#> 2019-10-09 12:00:00 "0.25"           
#> 2019-10-10 12:00:00 "0.26"           
#> 2019-10-11 12:00:00 "0.27"           
#> 2019-10-12 12:00:00 "0.27"           
#> 2019-10-13 12:00:00 "0.26"           
#> 2019-10-14 12:00:00 "0.27"           
#> 2019-10-15 12:00:00 "0.28"           
#> 2019-10-16 12:00:00 "0.27"           
#> 2019-10-17 12:00:00 "0.27"           
#> 2019-10-18 12:00:00 "0.28"           
#> 2019-10-19 12:00:00 "0.27"           
#> 2019-10-20 12:00:00 "0.27"           
#> 2019-10-21 12:00:00 "0.27"           
#> 2019-10-22 12:00:00 "0.26"           
#> 2019-10-23 12:00:00 "0.27"           
#> 2019-10-24 12:00:00 "0.26"           
#> 2019-10-25 12:00:00 "0.27"           
#> 2019-10-26 12:00:00 "0.28"           
#> 2019-10-27 12:00:00 "0.28"           
#> 2019-10-28 12:00:00 "0.26"           
#> 2019-10-29 12:00:00 "0.26"           
#> 2019-10-30 12:00:00 "0.26"           
#> 2019-10-31 12:00:00 "0.25"           
#> 2019-11-01 12:00:00 "0.26"           
#> 2019-11-02 12:00:00 "0.27"           
#> 2019-11-03 12:00:00 "0.25"           
#> 2019-11-04 12:00:00 "0.26"           
#> 2019-11-05 12:00:00 "0.25"           
#> 2019-11-06 12:00:00 "0.27"           
#> 2019-11-07 12:00:00 "0.26"           
#> 2019-11-08 12:00:00 "0.26"           
#> 2019-11-09 12:00:00 "0.26"           
#> 2019-11-10 12:00:00 "0.25"           
#> 2019-11-11 12:00:00 "0.25"           
#> 2019-11-12 12:00:00 "0.25"           
#> 2019-11-13 12:00:00 "0.27"           
#> 2019-11-14 12:00:00 "0.27"           
#> 2019-11-15 12:00:00 "0.28"           
#> 2019-11-16 12:00:00 "0.27"           
#> 2019-11-17 12:00:00 "0.25"           
#> 2019-11-18 12:00:00 "0.25"           
#> 2019-11-19 12:00:00 "0.24"           
#> 2019-11-20 12:00:00 "0.25"           
#> 2019-11-21 12:00:00 "0.25"           
#> 2019-11-22 12:00:00 "0.25"           
#> 2019-11-23 12:00:00 "0.25"           
#> 2019-11-24 12:00:00 "0.25"           
#> 2019-11-25 12:00:00 "0.24"           
#> 2019-11-26 12:00:00 "0.25"           
#> 2019-11-27 12:00:00 "0.26"           
#> 2019-11-28 12:00:00 "0.25"           
#> 2019-11-29 12:00:00 "0.25"           
#> 2019-11-30 12:00:00 "0.26"           
#> 2019-12-01 12:00:00 "0.26"           
#> 2019-12-02 12:00:00 "0.26"           
#> 2019-12-03 12:00:00 "0.26"           
#> 2019-12-04 12:00:00 "0.26"           
#> 2019-12-05 12:00:00 "0.27"           
#> 2019-12-06 12:00:00 "0.27"           
#> 2019-12-07 12:00:00 "0.28"           
#> 2019-12-08 12:00:00 "0.26"           
#> 2019-12-09 12:00:00 "0.27"           
#> 2019-12-10 12:00:00 "0.27"           
#> 2019-12-11 12:00:00 "0.27"           
#> 2019-12-12 12:00:00 "0.28"           
#> 2019-12-13 12:00:00 "0.29"           
#> 2019-12-14 12:00:00 "0.28"           
#> 2019-12-15 12:00:00 "0.29"           
#> 2019-12-16 12:00:00 "0.31"           
#> 2019-12-17 12:00:00 "0.31"           
#> 2019-12-18 12:00:00 "0.30"           
#> 2019-12-19 12:00:00 "0.28"           
#> 2019-12-20 12:00:00 "0.28"           
#> 2019-12-21 12:00:00 "0.27"           
#> 2019-12-22 12:00:00 "0.26"           
#> 2019-12-23 12:00:00 "0.27"           
#> 
#> $contract
#> [1] "What will be the balance of power in Congress after the 2020 election"
#>