Get bids and asks for a specific PredictIt market

Wrapper function to get data for a specific market.

single_market(id)

Arguments

id
Numerical code pertaining to the market. You can find a market's numerical code by consulting its URL or by first calling the all markets API.

Examples

markets <- all_markets()
id <- markets$id[1]
single_market(id)

#> # A tibble: 4 x 20
#>      id name  shortName image url   timeStamp           status contract_id
#>   <int> <chr> <chr>     <chr> <chr> <dttm>              <chr>        <int>
#> 1  2721 Whic… Which pa… http… http… 2019-12-24 13:03:13 Open          4390
#> 2  2721 Whic… Which pa… http… http… 2019-12-24 13:03:13 Open          4389
#> 3  2721 Whic… Which pa… http… http… 2019-12-24 13:03:13 Open          4388
#> 4  2721 Whic… Which pa… http… http… 2019-12-24 13:03:13 Open          4391
#> # … with 12 more variables: dateEnd <chr>, contract_image <chr>,
#> #   contract_name <chr>, contract_shortName <chr>, contract_status <chr>,
#> #   lastTradePrice <dbl>, bestBuyYesCost <dbl>, bestBuyNoCost <dbl>,
#> #   bestSellYesCost <dbl>, bestSellNoCost <dbl>, lastClosePrice <dbl>,
#> #   displayOrder <int>